Rename pyfaf rpm module
